change name to short_name; add trainee notes field; move js to dedicated file
